third_party/WebKit/LayoutTests/inspector/sources/debugger-pause/debugger-mute-exception-expected.txt

Tests that pause on exception is muted when conditional breakpoint is set to "false".

Script source was shown.
Script execution paused.
Call stack:
    0) throwAnException (debugger-mute-exception.html:18)
    1) handleClick (debugger-mute-exception.html:25)
    [setTimeout]
    0)  (:1)
Paused on exceptionTypeError: Cannot set property 'bar' of undefined
Script execution resumed.
